refactor: Extract Flow Scanner rules logic into shared module
- Create flow-scanner-rules.js with shared rule configuration logic
- Move getFlowScannerRules, normalizeSeverity, and related utilities
- Update imports in flow-scanner.js and options.js
- Improve code organization and reduce duplication
- Add error handling for localStorage retrieval
- Optimize rule lookup with Map
